All blue requirements functions work
