+++ code/src/components/TaskList.js
@@ -1,9 +1,10 @@
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector} from 'react-redux';
 import styled from 'styled-components';
-import { AddTask } from './AddTask';
 
+import { AddTask } from './AddTask';
 import { Task } from './Task';
+import { RemoveAll } from './RemoveAll';
 
 export const TaskList = () => {
 
@@ -11,14 +12,8 @@ export const TaskList = () => {
   const tasks = useSelector((store) => store.tasks);
   const tasksTodo = tasks.items.filter((task) => task.complete === false)
 
-  const dispatch = useDispatch();
-
-  const handleRemoveAllButton = () => {
-    dispatch(tasks.actions.removeAll())
-  };
-
   return (
-    <div className='list'>
+    <div>
       <CounterText>{tasksTodo.length}/{tasks.items.length} tasks uncompleted</CounterText>
       <AddTask />
       <ListofTasks>
@@ -26,8 +21,7 @@ export const TaskList = () => {
           <Task key={task.id} task={task} />
         ))}
       </ListofTasks>
-
-      <RemoveAllButton onClick={handleRemoveAllButton}><Emoji role="img" aria-label="Bin">🗑</Emoji> Remove all tasks!</RemoveAllButton>
+      <RemoveAll />
     </div>
   );
 };
@@ -38,19 +32,4 @@ const CounterText = styled.p`
 `
 const ListofTasks = styled.ul`
   padding: 0;
-`
-const RemoveAllButton = styled.button`
-  background: none;
-  border: 1px solid #869D7A;
-  border-radius: 4px;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  font-family: Segoe UI;
-  font-weight: bold;
-`
-const Emoji = styled.span`
-  font-size: 28px;
-  color: #6e6e6e;
-  padding-right: 3px;
 `
