+++ code/src/reducers/tasks.js
@@ -3,16 +3,16 @@ import { createSlice } from '@reduxjs/toolkit';
 export const tasks = createSlice({
   name: 'tasks',
   initialState: {
-    items: [
-      { id: 1, text: 'do this', complete: true, category: 'School' },
-      { id: 2, text: 'and do this', complete: false, category: 'Private' },
-      { id: 3, text: 'and then this too', complete: false, category: 'School' },
-    ],
+    items: [],
   }, 
   reducers: {
     addTask: (state, action) => {
-      const { id, text, category } = action.payload
-      state.items.push({ id, text, category })
+      const { text, category } = action.payload
+      state.items.push({ 
+        id: Date.now(), 
+        text, 
+        complete: false,
+        category })
     },
     removeTask: (state, action) => {
       state.items = state.items.filter((item) => item.id !== action.payload)
